Forecast Service/requirements.txt
quixstreams==2.3.1
python-dotenv
scikit-learn
numpy
